+++ test/agentchat/realtime_agent/test_swarm_start.py
@@ -16,7 +16,7 @@ from pytest import FixtureRequest
 from autogen import ConversableAgent
 from autogen.agentchat.realtime.experimental import RealtimeAgent, RealtimeObserver, WebSocketAudioAdapter
 from autogen.agentchat.realtime.experimental.realtime_swarm import register_swarm
-from autogen.tools.field import Field as AG2Field
+from autogen.tools.dependency_injection import Field as AG2Field
 
 from ...conftest import Credentials
 from .realtime_test_utils import text_to_speech, trace
